requirements.txt
aiohttp
sqlalchemy[asyncio]
aiohttp-sqlalchemy
asyncpg
alembic
PyYAML
black
